[2/3] Refactor dependency layer code
Reinstate stage_layer.py script
