+++ src/azul/lambda_layer.py
@@ -1,9 +1,7 @@
-from argparse import ArgumentParser
 import logging
 from pathlib import Path
 import shutil
 import subprocess
-import sys
 from zipfile import (
     ZipFile,
     ZipInfo,
@@ -16,7 +14,6 @@ from azul import (
     config,
 )
 from azul.files import file_sha1
-from azul.logging import configure_script_logging
 
 log = logging.getLogger(__name__)
 
@@ -83,20 +80,3 @@ class LayerBuilder:
     @cachedproperty
     def object_key(self):
         return config.layer_object_key(file_sha1(Path(config.project_root) / 'requirements.txt'))
-
-
-def main(argv):
-    parser = ArgumentParser(description='Package the requirements layer')
-    parser.add_argument('-f', '--force', action='store_true', help='force build and re-upload of layer. Also taint '
-                                                                   'Terraform resource to ensure it is re-provisioned')
-    options = parser.parse_args(argv)
-    packager = LayerBuilder()
-    if options.force:
-        packager.update_layer(force=True)
-    else:
-        packager.update_layer_if_necessary()
-
-
-if __name__ == '__main__':
-    configure_script_logging(log)
-    main(sys.argv[1:])
